+++ src/main/java/no/vegvesen/nvdbapi/client/clients/DatakatalogUpdateCallback.java
@@ -1,0 +1,6 @@
+package no.vegvesen.nvdbapi.client.clients;
+
+@FunctionalInterface
+public interface DatakatalogUpdateCallback {
+    void onDatakatalogUpdate();
+}